requirements.txt
python-dotenv
pymongo[srv]==3.12
pymongo
certifi
pandas
numpy
scikit-learn
dill
pyaml
mlflow
fastapi
uvicorn

#-e .